+++ egui_demo_lib/benches/benchmark.rs
@@ -36,11 +36,9 @@ pub fn criterion_benchmark(c: &mut Criterion) {
         let mut demo_windows = egui_demo_lib::DemoWindows::default();
         ctx.begin_frame(raw_input.clone());
         demo_windows.ui(&ctx);
-        let (_, paint_commands) = ctx.end_frame();
+        let (_, shapes) = ctx.end_frame();
 
-        c.bench_function("tessellate", |b| {
-            b.iter(|| ctx.tessellate(paint_commands.clone()))
-        });
+        c.bench_function("tessellate", |b| b.iter(|| ctx.tessellate(shapes.clone())));
     }
 
     {